GPX-698: Refactor multena-proxy for better readability and modularity
Refactored the multena-proxy code to improve its readability and maintainability. A large chunk of App and Config structs, previously part of types.go and main.go, were moved into a new file named config.go. The aim of the refactoring is to have a cleaner separation of the configuration handling from the main business logic. This further allows easier debuggability and can enhance the productivity of the developers, by providing a clear understanding and better navigation of the codebase.