GPX-698: Updated Multena Proxy README documentation, sample configs and architecture diagrams
Made several modifications to Multena Proxy's README.md, "config.yaml" and "labels.yaml" files to improve the clarity of existing instructions and provide more detailed explanation of the system. Changes include adjustments to the architecture diagram in the 'How does it work?' section, modifications to the 'request processing flow', and updates to the example 'labels.yaml' and 'config.yaml' files. These changes were necessary to improve the readability of the documentation, and keep them in line with the latest version of the software after recent updates to its functionality.